Download the input video from box
Small video is probably big enough to go in box not github.

The notebook does have aspects that are specific to video angle
etc. Also the training video affects what is recognized.

This video looks nice and it turns out that short is good,
but it lacks variety (nice and simple but less featureful).